chore:
re-structure the codea and add root route
